+++ cmd/backup.go
@@ -93,7 +93,7 @@ func RunBackup(ctx context.Context, flagMap map[string]any) error {
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),
-		pathrotation.NewPathRotator(),
+		pathrotation.NewPathRotator(time.Local),
 		pathretention.NewPathRetainer(
 			runConfig.Engine.Performance.DeleteWorkers,
 		),