+++ cmd/prune.go
@@ -97,7 +97,7 @@ func RunPrune(ctx context.Context, flagMap map[string]any) error {
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),
-		pathrotation.NewPathRotator(),
+		pathrotation.NewPathRotator(time.Local),
 		pathretention.NewPathRetainer(
 			runConfig.Engine.Performance.DeleteWorkers,
 		),