+++ cmd/restore.go
@@ -89,7 +89,7 @@ func RunRestore(ctx context.Context, flagMap map[string]any) error {
 			runConfig.Engine.Performance.SyncWorkers,
 			runConfig.Engine.Performance.MirrorWorkers,
 		),
-		pathrotation.NewPathRotator(),
+		pathrotation.NewPathRotator(time.Local),
 		pathretention.NewPathRetainer(
 			runConfig.Engine.Performance.DeleteWorkers,
 		),